requirements.txt
streamlit
langchain
python-dotenv
langchain-google-genai
langchain-core
langchain-community
faiss-cpu